+++ package.json
@@ -20,10 +20,9 @@
         "url": "https://github.com/coveo/coveo.analytics.js.git"
     },
     "license": "MIT",
-    "peerDependencies": {
-        "@react-native-async-storage/async-storage": "^1.13.2"
-    },
     "devDependencies": {
+        "react": "^17.0.1",
+        "react-native": "^0.63.3",
         "@react-native-async-storage/async-storage": "^1.13.2",
         "@types/express": "^4.17.2",
         "@types/fetch-mock": "^7.3.2",
