+++ src/react-native/index.ts
@@ -1,0 +1,2 @@
+export {ReactNativeRuntime} from './react-native-runtime';
+export {ReactNativeStorage} from './react-native-storage';
